Remove conf package and dependency on packethost/pkg/env:
Reorganize code to be able to remove the packethost/pkg/env
dependency. The idea is to follow up with a refactor
to clean up the cmd/boot/ package to handle cli flags
and env vars in a uniform way.
